# docs/openapi.json
{
  "openapi": "3.1.0",
...
  "info": {
    "title": "Payment Operations Investigator — Public Java API",
    "version": "0.1.0",
    "description": "Source-derived local synthetic portfolio API. Session-bound tenant/role authorization, immutable investigation evidence and human review; no real-payment commands. Only public Java routes are documented, not internal Python worker endpoints. Request DTOs reject unknown fields. Response metadata that Java deliberately leaves extensible remains optional. Monetary values use exact JSON-safe integer INR minor units. HTTP 200 is used for successful login, investigation and decision creation. This document does not claim production authentication, live-model success or database-administrator tamper resistance. Implements bounded original-synthetic OBPM NEFT ECA evidence imports; no Oracle source connector or banking-side commands. Import-specific validation and error status differ from ordinary DTO routes. A separate local mock inquiry transport now fetches original synthetic ECA evidence; it is not an Oracle product connector."
  },
  "servers": [
    {
      "url": "http://127.0.0.1:8088",
      "description": "Local Java API; web deployments normally proxy /api on the same origin."
    }
  ],
  "security": [
    {
      "SessionCookie": []
    }
  ],
...
  "paths": {
...
    "/api/cases": {
      "get": {
        "operationId": "listCases",
        "tags": [
          "Cases"
        ],
        "summary": "List visible synthetic cases",
        "description": "Tenant is derived from the authenticated session. Results are ordered by case ID; no pagination. Summaries omit events/ledgerEntries/webhooks/provider but include reconciliation. Unknown filter values return an empty list rather than validation errors.",
        "x-roles": [
          "ANALYST",
          "REVIEWER",
          "VIEWER"
        ],
        "parameters": [
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Case-insensitive trimmed substring across case ID, payment ID, title and merchant."
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Exact case-sensitive status string; omitted/blank means no filter."
          },
          {
            "name": "priority",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Exact case-sensitive priority string; omitted/blank means no filter."
          }
        ],
        "responses": {
          "200": {
            "description": "Success.",
            "headers": {
              "X-Request-Id": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CaseList"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthenticated"
          },
          "422": {
            "$ref": "#/components/responses/InvalidData"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
    "/api/cases/{id}": {
      "get": {
        "operationId": "getCase",
        "tags": [
          "Cases"
        ],
        "summary": "Read a tenant-scoped case and monetary facts",
        "description": "",
        "x-roles": [
          "ANALYST",
          "REVIEWER",
          "VIEWER"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/CaseId"
          }
        ],
        "responses": {
          "200": {
            "description": "Success.",
            "headers": {
              "X-Request-Id": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CaseDetail"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthenticated"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "422": {
            "$ref": "#/components/responses/InvalidData"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
    "/api/cases/{id}/investigations": {
      "post": {
        "operationId": "createInvestigation",
        "tags": [
          "Investigations"
        ],
        "summary": "Create an immutable evidence-linked proposal",
        "description": "Synchronous real worker call; success is HTTP 200, not 201/202. Java validates result evidence/schema, stamps identity/time/snapshot hash, stores the proposal and audit atomically, and changes the case to AWAITING_REVIEW with version+1. Before accepting RESOLVE_CASE, Java independently requires its authorized snapshot's valid money, available integer provider payout and known zero discrepancy; invalid resolution output returns 503 INVALID_WORKER_RESPONSE. The worker returns a valid INSUFFICIENT_EVIDENCE/REQUEST_EVIDENCE proposal for an otherwise supported timeout, duplicate-delivery or ordering case whose payout/reconciliation is unknown or nonzero. A sole supported confirmed missing-refund exception retains MISSING_REFUND/ESCALATE. Concurrent case changes reject the stale result. No client idempotency key is supported on this endpoint; repeated requests create separate investigations. The configured worker request deadline defaults/caps at 390 seconds. Failure records an INVESTIGATION_FAILED audit event without fabricating a proposal or changing case status/version; timeout does not prove downstream provider computation was cancelled. Current worker assessment fields (outcome, summary, confidence, missingEvidence and proposal) come from deterministic evidence rules. Ollama performs tool planning and typed selection of one or two authorized catalog facts for supported evidence. The service supplies exact selected fact sentences and their complete evidence/policy links, without adding unselected claims or repairing prose. Insufficient-evidence results deliberately skip the finding stage and return findings=[] after actual tool planning. Optional metrics record selection, wording and assessment origins plus catalog version/hash/selected IDs. Historical model-written findings remain unchanged. Stored proposals still require independent review. New Java results also persist the exact input caseSnapshot. OBPM_NEFT selects its separate ECA tools/rules and scoped policy catalog: only OBPM_ECA_TIMEOUT/INSUFFICIENT_EVIDENCE with REQUEST_EVIDENCE/ESCALATE are accepted; RESOLVE_CASE is rejected regardless of model or reconciliation fields. OBPM live planning returns a typed permutation of all four mandatory evidence tools; the model orders collection rather than deciding which required sources to omit. Invalid/incomplete plans fail, with no automatic completion, retry or replay fallback. Optional toolPlanningScope and orderedToolNames record this distinction.",
        "x-roles": [
          "ANALYST",
          "REVIEWER"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/CaseId"
          },
          {
            "$ref": "#/components/parameters/CsrfToken"
          }
        ],
        "responses": {
          "200": {
            "description": "Success.",
            "headers": {
              "X-Request-Id": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Investigation"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthenticated"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "422": {
            "$ref": "#/components/responses/InvalidData"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          },
          "503": {
            "$ref": "#/components/responses/WorkerUnavailable"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InvestigateRequest"
              }
            }
          }
        }
      },
      "get": {
        "operationId": "listInvestigations",
        "tags": [
          "Investigations"
        ],
        "summary": "List immutable proposals for a case",
        "description": "Newest createdAt first, then descending ID. Stored proposal status remains AWAITING_REVIEW after review; obtain current case state and exported decisions separately.",
        "x-roles": [
          "ANALYST",
          "REVIEWER",
          "VIEWER"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/CaseId"
          }
        ],
        "responses": {
          "200": {
            "description": "Success.",
            "headers": {
              "X-Request-Id": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvestigationList"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthenticated"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "422": {
            "$ref": "#/components/responses/InvalidData"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
    "/api/investigations/{id}": {
      "get": {
        "operationId": "getInvestigation",
        "tags": [
          "Investigations"
        ],
        "summary": "Read a stored tenant-scoped investigation",
        "description": "",
        "x-roles": [
          "ANALYST",
          "REVIEWER",
          "VIEWER"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/InvestigationId"
          }
        ],
        "responses": {
          "200": {
            "description": "Success.",
            "headers": {
              "X-Request-Id": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Investigation"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthenticated"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
    "/api/cases/{id}/decisions": {
      "post": {
        "operationId": "reviewInvestigation",
        "tags": [
          "Reviews"
        ],
        "summary": "Approve or reject a stored case proposal once",
        "description": "Reviewer must differ from investigation creator. The action comes only from the stored proposal; unknown request fields including action are rejected. Case and proposal versions must match expectedVersion. Exact command replay with the same reviewer/tenant/key is checked before version rejection and returns 200/replayed=true with the original decision ID/version, adding no audit/state change. A new decision atomically records reviewer/note/time, increments case version and appends audit. No payment or ledger values are mutated. An imported evidence refresh increments the case version and makes older proposals stale. OBPM proposals cannot resolve a case; approval only changes the local case workflow.",
        "x-roles": [
          "REVIEWER"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/CaseId"
          },
          {
            "$ref": "#/components/parameters/CsrfToken"
          },
          {
            "$ref": "#/components/parameters/IdempotencyKey"
          }
        ],
        "responses": {
          "200": {
            "description": "Success.",
            "headers": {
              "X-Request-Id": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewDecision"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthenticated"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "422": {
            "$ref": "#/components/responses/InvalidData"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DecisionRequest"
              }
            }
          }
        }
      }
    },
    "/api/cases/{id}/audit": {
      "get": {
        "operationId": "listAudit",
        "tags": [
          "Reviews"
        ],
        "summary": "Read append-only application audit for a case",
        "description": "Chronological occurredAt order, then ID. Application code provides no update/delete endpoint. This is not a claim of database-administrator tamper resistance.",
        "x-roles": [
          "ANALYST",
          "REVIEWER",
          "VIEWER"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/CaseId"
          }
        ],
        "responses": {
          "200": {
            "description": "Success.",
            "headers": {
              "X-Request-Id": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuditList"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthenticated"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "422": {
            "$ref": "#/components/responses/InvalidData"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
    "/api/cases/{id}/export": {
      "get": {
        "operationId": "exportCase",
        "tags": [
          "Reviews"
        ],
        "summary": "Export an authorized evidence packet",
        "description": "Returns JSON in the response body, without a guaranteed attachment filename. Includes case, immutable investigations, sanitized review decisions and audit; omits internal review command hashes/keys. Also includes evidenceVersions with full immutable OBPM snapshot payloads and saved per-investigation caseSnapshot when present; never substitutes current evidence for an older saved investigation.",
        "x-roles": [
          "ANALYST",
          "REVIEWER",
          "VIEWER"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/CaseId"
          }
        ],
        "responses": {
          "200": {
            "description": "Success.",
            "headers": {
              "X-Request-Id": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Export"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthenticated"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "422": {
            "$ref": "#/components/responses/InvalidData"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
    "/api/dashboard": {
      "get": {
        "operationId": "dashboard",
        "tags": [
          "System"
        ],
        "summary": "Read tenant totals and recent activity",
        "description": "Counts and total cover all visible cases. Java uses checked integer arithmetic and rejects an unsafe aggregate rather than returning a rounded or wrapped amount. These are imported-case metrics, not all-payment OBPM analytics or success rates.",
        "x-roles": [
          "ANALYST",
          "REVIEWER",
          "VIEWER"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Success.",
            "headers": {
              "X-Request-Id": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Dashboard"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthenticated"
          },
          "422": {
            "$ref": "#/components/responses/InvalidData"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
    "/api/knowledge": {
      "get": {
        "operationId": "knowledge",
        "tags": [
          "Knowledge"
        ],
        "summary": "Read tenant-visible original runbooks",
        "description": "Authenticated proxy to the worker. Java filters tenantId to the current tenant or shared *. Library may include expired/historical versions; it is not the date-filtered retrieval result.",
        "x-roles": [
          "ANALYST",
          "REVIEWER",
          "VIEWER"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Success.",
            "headers": {
              "X-Request-Id": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KnowledgeList"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthenticated"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          },
          "503": {
            "$ref": "#/components/responses/WorkerUnavailable"
          }
        }
      }
    },
    "/api/system": {
      "get": {
        "operationId": "system",
        "tags": [
          "System"
        ],
        "summary": "Read dataset and dependency availability",
        "description": "Worker/knowledge failure is represented in a successful response: workerStatus may be DOWN, and knowledgeAvailable=false makes runbookCount unavailable despite its numeric zero placeholder. supportedModes lists implemented modes and does not assert an installed model, successful inference, or semantic retrieval availability.",
        "x-roles": [
          "ANALYST",
          "REVIEWER",
          "VIEWER"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Success.",
            "headers": {
              "X-Request-Id": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/System"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthenticated"
          },
          "422": {
            "$ref": "#/components/responses/InvalidData"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "SessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "POI_SESSION",
        "description": "Opaque HttpOnly SameSite=Strict cookie obtained via POST /api/auth/login. Authenticated mutations also require the separate X-CSRF-Token header. Do not send tenant/role headers as authorization."
      }
    },
    "parameters": {
      "CaseId": {
        "name": "id",
        "in": "path",
        "required": true,
        "description": "Case ID within the authenticated tenant.",
        "schema": {
          "type": "string"
        }
      },
      "InvestigationId": {
        "name": "id",
        "in": "path",
        "required": true,
        "description": "Immutable investigation ID within the authenticated tenant.",
        "schema": {
          "type": "string"
        }
      },
      "CsrfToken": {
        "name": "X-CSRF-Token",
        "in": "header",
        "required": true,
        "description": "Token returned by login or /api/auth/me, bound to the current session. Required for all protected POST operations.",
        "schema": {
          "type": "string",
          "minLength": 1,
          "pattern": "\\S"
        }
      },
      "IdempotencyKey": {
        "name": "Idempotency-Key",
        "in": "header",
        "required": true,
        "description": "Unique review command key scoped to authenticated tenant and reviewer. Exact same body/key returns the original decision; a changed body with the same key returns 409.",
        "schema": {
          "type": "string",
          "minLength": 8,
          "maxLength": 200,
          "pattern": "^[A-Za-z0-9._:-]{8,200}$"
        }
      }
    },
    "responses": {
      "BadRequest": {
        "description": "400: malformed JSON, unknown request fields, invalid validation fields/mode/decision, or missing/invalid Idempotency-Key.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        },
        "headers": {
          "X-Request-Id": {
            "schema": {
              "type": "string"
            }
          }
        }
      },
      "Unauthenticated": {
        "description": "401: no valid authenticated session, or invalid login credentials.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        },
        "headers": {
          "X-Request-Id": {
            "schema": {
              "type": "string"
            }
          }
        }
      },
      "Forbidden": {
        "description": "403: missing/invalid CSRF, disallowed role, maker-checker violation, or cross-site login.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        },
        "headers": {
          "X-Request-Id": {
            "schema": {
              "type": "string"
            }
          }
        }
      },
      "NotFound": {
        "description": "404: record absent or outside the authenticated tenant; cross-tenant existence is not disclosed.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        },
        "headers": {
          "X-Request-Id": {
            "schema": {
              "type": "string"
            }
          }
        }
      },
      "Conflict": {
        "description": "409: stale/superseded proposal or case version, already-reviewed proposal, changed payload under the same idempotency key, or database concurrency conflict.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        },
        "headers": {
          "X-Request-Id": {
            "schema": {
              "type": "string"
            }
          }
        }
      },
      "InvalidData": {
        "description": "422: unsupported/invalid stored monetary data, mixed currencies, totals outside the JSON-safe range, or invalid stored proposal. No partial monetary result is returned.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        },
        "headers": {
          "X-Request-Id": {
            "schema": {
              "type": "string"
            }
          }
        }
      },
      "WorkerUnavailable": {
        "description": "503: worker/provider unavailable, timeout, non-200 worker response, or invalid worker schema/evidence. INVALID_WORKER_RESPONSE also covers a RESOLVE_CASE proposal without Java-authoritative valid money, available integer payout and known zero discrepancy. Worker errors are translated to 503; no replay fallback is fabricated.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        },
        "headers": {
          "X-Request-Id": {
            "schema": {
              "type": "string"
            }
          }
        }
      },
      "InternalError": {
        "description": "500: unexpected server/framework error; response does not expose stack traces.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        },
        "headers": {
          "X-Request-Id": {
            "schema": {
              "type": "string"
            }
          }
        }
      },
...
    },
    "schemas": {
...
      "InvestigateRequest": {
        "type": "object",
        "properties": {
          "question": {
            "type": "string",
            "minLength": 1,
            "pattern": "\\S",
            "maxLength": 2000
          },
          "mode": {
            "$ref": "#/components/schemas/Mode"
          }
        },
        "required": [
          "question",
          "mode"
        ],
        "additionalProperties": false
      },
...
      "CaseList": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CaseSummary"
            }
          },
          "total": {
            "type": "integer",
            "minimum": 0
          }
        },
        "required": [
          "items",
          "total"
        ],
        "additionalProperties": false
      },
...
      "InvestigationList": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Investigation"
            }
          }
        },
        "required": [
          "items"
        ],
        "additionalProperties": false
      },
...
    }
  },
...
}